fix show progress percentage for big files
suggested by @fftmp

+++ heimdall/source/BridgeManager.cpp
@@ -1052,7 +1052,7 @@ bool BridgeManager::SendFile(FILE *file, unsigned int destination, unsigned int 
 			lastSequenceSize++;
 	}
 
-	unsigned int bytesTransferred = 0;
+	unsigned long bytesTransferred = 0;
 	unsigned int currentPercent;
 	unsigned int previousPercent = 0;
 	Interface::Print("0%%");
